backend/requirements.txt
cohere>=4.9.0
qdrant-client>=1.8.0
beautifulsoup4>=4.12.2
requests>=2.31.0
python-dotenv>=1.0.0
argparse>=1.4.0